cc contracts deployed and initialized

+++ contracts/storage/OrderBoxRelayerStorage.sol
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.20;
 
-import { OrderRelayerType } from "../library/OrderRelayerType.sol";
-
 contract OrderBoxRelayerStorage {
-    // chainId => orderSafe address
-    mapping(uint256 => address) public orderSafe;
+    // eid => orderSafe address
+    mapping(uint32 => address) public orderSafe;
     address public orderBox;
+
+    /* ========== Storage Slots + Gap == 50 ========== */
     uint256[50] private _gap;
 }
